Restrict user overrides for remote store related index settings (#8812)
