+++ server/src/internalClusterTest/java/org/opensearch/remotestore/RemoteIndexRecoveryIT.java
@@ -23,14 +23,20 @@ import org.opensearch.test.OpenSearchIntegTestCase;
 
 import java.nio.file.Path;
 
+import static org.opensearch.remotestore.RemoteStoreBaseIntegTestCase.remoteStoreClusterSettings;
 import static org.opensearch.test.hamcrest.OpenSearchAssertions.assertAcked;
 
 @OpenSearchIntegTestCase.ClusterScope(scope = OpenSearchIntegTestCase.Scope.TEST, numDataNodes = 0)
 public class RemoteIndexRecoveryIT extends IndexRecoveryIT {
 
-    protected static final String REPOSITORY_NAME = "test-remore-store-repo";
+    protected static final String REPOSITORY_NAME = "test-remote-store-repo";
 
     protected Path absolutePath;
+
+    @Override
+    protected Settings nodeSettings(int nodeOrdinal) {
+        return Settings.builder().put(super.nodeSettings(nodeOrdinal)).put(remoteStoreClusterSettings(REPOSITORY_NAME)).build();
+    }
 
     @Override
     protected Settings featureFlagSettings() {
@@ -57,9 +63,6 @@ public class RemoteIndexRecoveryIT extends IndexRecoveryIT {
         return Settings.builder()
             .put(super.indexSettings())
             .put(IndexModule.INDEX_QUERY_CACHE_ENABLED_SETTING.getKey(), false)
-            .put(IndexMetadata.SETTING_REMOTE_STORE_ENABLED, true)
-            .put(IndexMetadata.SETTING_REMOTE_SEGMENT_STORE_REPOSITORY, REPOSITORY_NAME)
-            .put(IndexMetadata.SETTING_REMOTE_TRANSLOG_STORE_REPOSITORY, REPOSITORY_NAME)
             .put(IndexSettings.INDEX_REFRESH_INTERVAL_SETTING.getKey(), "300s")
             .put(IndexMetadata.SETTING_REPLICATION_TYPE, ReplicationType.SEGMENT)
             .build();
@@ -81,7 +84,91 @@ public class RemoteIndexRecoveryIT extends IndexRecoveryIT {
     }
 
     @Override
-    protected boolean shouldAssertOngoingRecoveryInRerouteRecovery() {
-        return false;
+    public void testUsesFileBasedRecoveryIfRetentionLeaseMissing() {
+        // Retention lease based tests not applicable for remote store;
+    }
+
+    @Override
+    public void testPeerRecoveryTrimsLocalTranslog() {
+        // Peer recovery usecase not valid for remote enabled indices
+    }
+
+    @Override
+    public void testHistoryRetention() {
+        // History retention not applicable for remote store
+    }
+
+    @Override
+    public void testUsesFileBasedRecoveryIfOperationsBasedRecoveryWouldBeUnreasonable() {
+        // History retention not applicable for remote store
+    }
+
+    @Override
+    public void testUsesFileBasedRecoveryIfRetentionLeaseAheadOfGlobalCheckpoint() {
+        // History retention not applicable for remote store
+    }
+
+    @Override
+    public void testRecoverLocallyUpToGlobalCheckpoint() {
+        // History retention not applicable for remote store
+    }
+
+    @Override
+    public void testCancelNewShardRecoveryAndUsesExistingShardCopy() {
+        // History retention not applicable for remote store
+    }
+
+    @AwaitsFix(bugUrl = "https://github.com/opensearch-project/OpenSearch/issues/8919")
+    @Override
+    public void testReservesBytesDuringPeerRecoveryPhaseOne() {
+
+    }
+
+    @AwaitsFix(bugUrl = "https://github.com/opensearch-project/OpenSearch/issues/8919")
+    @Override
+    public void testAllocateEmptyPrimaryResetsGlobalCheckpoint() {
+
+    }
+
+    @AwaitsFix(bugUrl = "https://github.com/opensearch-project/OpenSearch/issues/8919")
+    @Override
+    public void testDoesNotCopyOperationsInSafeCommit() {
+
+    }
+
+    @AwaitsFix(bugUrl = "https://github.com/opensearch-project/OpenSearch/issues/8919")
+    @Override
+    public void testRepeatedRecovery() {
+
+    }
+
+    @AwaitsFix(bugUrl = "https://github.com/opensearch-project/OpenSearch/issues/8919")
+    @Override
+    public void testDisconnectsWhileRecovering() {
+
+    }
+
+    @AwaitsFix(bugUrl = "https://github.com/opensearch-project/OpenSearch/issues/8919")
+    @Override
+    public void testTransientErrorsDuringRecoveryAreRetried() {
+
+    }
+
+    @AwaitsFix(bugUrl = "https://github.com/opensearch-project/OpenSearch/issues/8919")
+    @Override
+    public void testDoNotInfinitelyWaitForMapping() {
+
+    }
+
+    @AwaitsFix(bugUrl = "https://github.com/opensearch-project/OpenSearch/issues/8919")
+    @Override
+    public void testDisconnectsDuringRecovery() {
+
+    }
+
+    @AwaitsFix(bugUrl = "https://github.com/opensearch-project/OpenSearch/issues/8919")
+    @Override
+    public void testReplicaRecovery() {
+
     }
 }
