+++ server/src/internalClusterTest/java/org/opensearch/remotestore/multipart/RemoteStoreMultipartFileCorruptionIT.java
@@ -8,31 +8,22 @@
 
 package org.opensearch.remotestore.multipart;
 
-import org.junit.After;
 import org.junit.Before;
-import org.opensearch.action.index.IndexResponse;
 import org.opensearch.action.support.IndicesOptions;
 import org.opensearch.cluster.metadata.IndexMetadata;
-import org.opensearch.common.UUIDs;
 import org.opensearch.common.settings.Settings;
-import org.opensearch.common.util.FeatureFlags;
 import org.opensearch.index.IndexModule;
 import org.opensearch.indices.replication.common.ReplicationType;
 import org.opensearch.plugins.Plugin;
-import org.opensearch.remotestore.multipart.mocks.MockFsRepository;
+import org.opensearch.remotestore.RemoteStoreBaseIntegTestCase;
 import org.opensearch.remotestore.multipart.mocks.MockFsRepositoryPlugin;
-import org.opensearch.test.OpenSearchIntegTestCase;
 
-import java.nio.file.Path;
 import java.util.Collection;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
-import static org.opensearch.test.hamcrest.OpenSearchAssertions.assertAcked;
+public class RemoteStoreMultipartFileCorruptionIT extends RemoteStoreBaseIntegTestCase {
 
-public class RemoteStoreMultipartFileCorruptionIT extends OpenSearchIntegTestCase {
-
-    protected static final String REPOSITORY_NAME = "test-remore-store-repo";
     private static final String INDEX_NAME = "remote-store-test-idx-1";
 
     @Override
@@ -40,34 +31,9 @@ public class RemoteStoreMultipartFileCorruptionIT extends OpenSearchIntegTestCas
         return Stream.concat(super.nodePlugins().stream(), Stream.of(MockFsRepositoryPlugin.class)).collect(Collectors.toList());
     }
 
-    @Override
-    protected Settings featureFlagSettings() {
-        return Settings.builder().put(super.featureFlagSettings()).put(FeatureFlags.REMOTE_STORE, "true").build();
-    }
-
     @Before
     public void setup() {
-        internalCluster().startClusterManagerOnlyNode();
-        Path absolutePath = randomRepoPath().toAbsolutePath();
-        putRepository(absolutePath);
-    }
-
-    protected void putRepository(Path path) {
-        assertAcked(
-            clusterAdmin().preparePutRepository(REPOSITORY_NAME)
-                .setType(MockFsRepositoryPlugin.TYPE)
-                .setSettings(
-                    Settings.builder()
-                        .put("location", path)
-                        // custom setting for MockFsRepositoryPlugin
-                        .put(MockFsRepository.TRIGGER_DATA_INTEGRITY_FAILURE.getKey(), true)
-                )
-        );
-    }
-
-    @After
-    public void teardown() {
-        assertAcked(clusterAdmin().prepareDeleteRepository(REPOSITORY_NAME));
+        setupRepo();
     }
 
     protected Settings remoteStoreIndexSettings() {
@@ -78,17 +44,7 @@ public class RemoteStoreMultipartFileCorruptionIT extends OpenSearchIntegTestCas
             .put(IndexMetadata.SETTING_NUMBER_OF_REPLICAS, 0)
             .put(IndexModule.INDEX_QUERY_CACHE_ENABLED_SETTING.getKey(), false)
             .put(IndexMetadata.SETTING_REPLICATION_TYPE, ReplicationType.SEGMENT)
-            .put(IndexMetadata.SETTING_REMOTE_STORE_ENABLED, true)
-            .put(IndexMetadata.SETTING_REMOTE_SEGMENT_STORE_REPOSITORY, REPOSITORY_NAME)
-            .put(IndexMetadata.SETTING_REMOTE_TRANSLOG_STORE_REPOSITORY, REPOSITORY_NAME)
             .build();
-    }
-
-    private IndexResponse indexSingleDoc() {
-        return client().prepareIndex(INDEX_NAME)
-            .setId(UUIDs.randomBase64UUID())
-            .setSource(randomAlphaOfLength(5), randomAlphaOfLength(5))
-            .get();
     }
 
     public void testLocalFileCorruptionDuringUpload() {
@@ -97,7 +53,7 @@ public class RemoteStoreMultipartFileCorruptionIT extends OpenSearchIntegTestCas
         ensureYellowAndNoInitializingShards(INDEX_NAME);
         ensureGreen(INDEX_NAME);
 
-        indexSingleDoc();
+        indexSingleDoc(INDEX_NAME);
 
         client().admin()
             .indices()
