+++ server/src/internalClusterTest/java/org/opensearch/snapshots/DeleteSnapshotIT.java
@@ -25,15 +25,18 @@ import java.util.Locale;
 import java.util.stream.Stream;
 
 import static org.hamcrest.Matchers.is;
+import static org.opensearch.remotestore.RemoteStoreBaseIntegTestCase.remoteStoreClusterSettings;
 import static org.opensearch.test.hamcrest.OpenSearchAssertions.assertAcked;
 
 @OpenSearchIntegTestCase.ClusterScope(scope = OpenSearchIntegTestCase.Scope.TEST, numDataNodes = 0)
 public class DeleteSnapshotIT extends AbstractSnapshotIntegTestCase {
 
+    private static final String REMOTE_REPO_NAME = "remote-store-repo-name";
+
     public void testDeleteSnapshot() throws Exception {
         disableRepoConsistencyCheck("Remote store repository is being used in the test");
         FeatureFlagSetter.set(FeatureFlags.REMOTE_STORE);
-        internalCluster().startClusterManagerOnlyNode();
+        internalCluster().startClusterManagerOnlyNode(remoteStoreClusterSettings(REMOTE_REPO_NAME));
         internalCluster().startDataOnlyNode();
 
         final String snapshotRepoName = "snapshot-repo-name";
@@ -41,20 +44,19 @@ public class DeleteSnapshotIT extends AbstractSnapshotIntegTestCase {
         createRepository(snapshotRepoName, "fs", snapshotRepoPath);
 
         final Path remoteStoreRepoPath = randomRepoPath();
-        final String remoteStoreRepoName = "remote-store-repo-name";
-        createRepository(remoteStoreRepoName, "fs", remoteStoreRepoPath);
+        createRepository(REMOTE_REPO_NAME, "fs", remoteStoreRepoPath);
 
         final String indexName = "index-1";
         createIndexWithRandomDocs(indexName, randomIntBetween(5, 10));
 
         final String remoteStoreEnabledIndexName = "remote-index-1";
-        final Settings remoteStoreEnabledIndexSettings = getRemoteStoreBackedIndexSettings(remoteStoreRepoName);
+        final Settings remoteStoreEnabledIndexSettings = getRemoteStoreBackedIndexSettings();
         createIndex(remoteStoreEnabledIndexName, remoteStoreEnabledIndexSettings);
         indexRandomDocs(remoteStoreEnabledIndexName, randomIntBetween(5, 10));
 
         final String snapshot = "snapshot";
         createFullSnapshot(snapshotRepoName, snapshot);
-        assert (getLockFilesInRemoteStore(remoteStoreEnabledIndexName, remoteStoreRepoName).length == 0);
+        assert (getLockFilesInRemoteStore(remoteStoreEnabledIndexName, REMOTE_REPO_NAME).length == 0);
         assert (getRepositoryData(snapshotRepoName).getSnapshotIds().size() == 1);
 
         assertAcked(startDeleteSnapshot(snapshotRepoName, snapshot).get());
@@ -64,32 +66,31 @@ public class DeleteSnapshotIT extends AbstractSnapshotIntegTestCase {
     public void testDeleteShallowCopySnapshot() throws Exception {
         disableRepoConsistencyCheck("Remote store repository is being used in the test");
         FeatureFlagSetter.set(FeatureFlags.REMOTE_STORE);
-        internalCluster().startClusterManagerOnlyNode();
+        internalCluster().startClusterManagerOnlyNode(remoteStoreClusterSettings(REMOTE_REPO_NAME));
         internalCluster().startDataOnlyNode();
 
         final String snapshotRepoName = "snapshot-repo-name";
         createRepository(snapshotRepoName, "fs", snapshotRepoSettingsForShallowCopy());
 
         final Path remoteStoreRepoPath = randomRepoPath();
-        final String remoteStoreRepoName = "remote-store-repo-name";
-        createRepository(remoteStoreRepoName, "fs", remoteStoreRepoPath);
+        createRepository(REMOTE_REPO_NAME, "fs", remoteStoreRepoPath);
 
         final String indexName = "index-1";
         createIndexWithRandomDocs(indexName, randomIntBetween(5, 10));
 
         final String remoteStoreEnabledIndexName = "remote-index-1";
-        final Settings remoteStoreEnabledIndexSettings = getRemoteStoreBackedIndexSettings(remoteStoreRepoName);
+        final Settings remoteStoreEnabledIndexSettings = getRemoteStoreBackedIndexSettings();
         createIndex(remoteStoreEnabledIndexName, remoteStoreEnabledIndexSettings);
         indexRandomDocs(remoteStoreEnabledIndexName, randomIntBetween(5, 10));
 
         final String shallowSnapshot = "shallow-snapshot";
         createFullSnapshot(snapshotRepoName, shallowSnapshot);
-        assert (getLockFilesInRemoteStore(remoteStoreEnabledIndexName, remoteStoreRepoName).length == 1);
+        assert (getLockFilesInRemoteStore(remoteStoreEnabledIndexName, REMOTE_REPO_NAME).length == 1);
         assert (getRepositoryData(snapshotRepoName).getSnapshotIds().size() == 1);
 
         assertAcked(startDeleteSnapshot(snapshotRepoName, shallowSnapshot).get());
         assert (getRepositoryData(snapshotRepoName).getSnapshotIds().size() == 0);
-        assert (getLockFilesInRemoteStore(remoteStoreEnabledIndexName, remoteStoreRepoName).length == 0);
+        assert (getLockFilesInRemoteStore(remoteStoreEnabledIndexName, REMOTE_REPO_NAME).length == 0);
     }
 
     // Deleting multiple shallow copy snapshots as part of single delete call with repo having only shallow copy snapshots.
@@ -97,10 +98,13 @@ public class DeleteSnapshotIT extends AbstractSnapshotIntegTestCase {
         disableRepoConsistencyCheck("Remote store repository is being used in the test");
         FeatureFlagSetter.set(FeatureFlags.REMOTE_STORE);
 
-        internalCluster().startClusterManagerOnlyNode();
+        internalCluster().startClusterManagerOnlyNode(remoteStoreClusterSettings(REMOTE_REPO_NAME));
         internalCluster().startDataOnlyNode();
         final Client clusterManagerClient = internalCluster().clusterManagerClient();
         ensureStableCluster(2);
+
+        final Path remoteStoreRepoPath = randomRepoPath();
+        createRepository(REMOTE_REPO_NAME, "fs", remoteStoreRepoPath);
 
         final String snapshotRepoName = "snapshot-repo-name";
         final Path snapshotRepoPath = randomRepoPath();
@@ -108,12 +112,8 @@ public class DeleteSnapshotIT extends AbstractSnapshotIntegTestCase {
         final String testIndex = "index-test";
         createIndexWithContent(testIndex);
 
-        final Path remoteStoreRepoPath = randomRepoPath();
-        final String remoteStoreRepoName = "remote-store-repo-name";
-        createRepository(remoteStoreRepoName, "fs", remoteStoreRepoPath);
-
         final String remoteStoreEnabledIndexName = "remote-index-1";
-        final Settings remoteStoreEnabledIndexSettings = getRemoteStoreBackedIndexSettings(remoteStoreRepoName);
+        final Settings remoteStoreEnabledIndexSettings = getRemoteStoreBackedIndexSettings();
         createIndex(remoteStoreEnabledIndexName, remoteStoreEnabledIndexSettings);
         indexRandomDocs(remoteStoreEnabledIndexName, randomIntBetween(5, 10));
 
@@ -122,7 +122,7 @@ public class DeleteSnapshotIT extends AbstractSnapshotIntegTestCase {
         List<String> shallowCopySnapshots = createNSnapshots(snapshotRepoName, totalShallowCopySnapshotsCount);
         List<String> snapshotsToBeDeleted = shallowCopySnapshots.subList(0, randomIntBetween(2, totalShallowCopySnapshotsCount));
         int tobeDeletedSnapshotsCount = snapshotsToBeDeleted.size();
-        assert (getLockFilesInRemoteStore(remoteStoreEnabledIndexName, remoteStoreRepoName).length == totalShallowCopySnapshotsCount);
+        assert (getLockFilesInRemoteStore(remoteStoreEnabledIndexName, REMOTE_REPO_NAME).length == totalShallowCopySnapshotsCount);
         assert (getRepositoryData(snapshotRepoName).getSnapshotIds().size() == totalShallowCopySnapshotsCount);
         // Deleting subset of shallow copy snapshots
         assertAcked(
@@ -132,7 +132,7 @@ public class DeleteSnapshotIT extends AbstractSnapshotIntegTestCase {
                 .get()
         );
         assert (getRepositoryData(snapshotRepoName).getSnapshotIds().size() == totalShallowCopySnapshotsCount - tobeDeletedSnapshotsCount);
-        assert (getLockFilesInRemoteStore(remoteStoreEnabledIndexName, remoteStoreRepoName).length == totalShallowCopySnapshotsCount
+        assert (getLockFilesInRemoteStore(remoteStoreEnabledIndexName, REMOTE_REPO_NAME).length == totalShallowCopySnapshotsCount
             - tobeDeletedSnapshotsCount);
     }
 
@@ -144,7 +144,7 @@ public class DeleteSnapshotIT extends AbstractSnapshotIntegTestCase {
         disableRepoConsistencyCheck("Remote store repository is being used in the test");
         FeatureFlagSetter.set(FeatureFlags.REMOTE_STORE);
 
-        internalCluster().startClusterManagerOnlyNode();
+        internalCluster().startClusterManagerOnlyNode(remoteStoreClusterSettings(REMOTE_REPO_NAME));
         final String dataNode = internalCluster().startDataOnlyNode();
         ensureStableCluster(2);
         final String clusterManagerNode = internalCluster().getClusterManagerName();
@@ -156,11 +156,10 @@ public class DeleteSnapshotIT extends AbstractSnapshotIntegTestCase {
         createIndexWithContent(testIndex);
 
         final Path remoteStoreRepoPath = randomRepoPath();
-        final String remoteStoreRepoName = "remote-store-repo-name";
-        createRepository(remoteStoreRepoName, "fs", remoteStoreRepoPath);
+        createRepository(REMOTE_REPO_NAME, "fs", remoteStoreRepoPath);
 
         final String remoteStoreEnabledIndexName = "remote-index-1";
-        final Settings remoteStoreEnabledIndexSettings = getRemoteStoreBackedIndexSettings(remoteStoreRepoName);
+        final Settings remoteStoreEnabledIndexSettings = getRemoteStoreBackedIndexSettings();
         createIndex(remoteStoreEnabledIndexName, remoteStoreEnabledIndexSettings);
         indexRandomDocs(remoteStoreEnabledIndexName, randomIntBetween(5, 10));
 
@@ -201,7 +200,7 @@ public class DeleteSnapshotIT extends AbstractSnapshotIntegTestCase {
 
         int totalSnapshotsCount = totalFullCopySnapshotsCount + totalShallowCopySnapshotsCount;
 
-        assert (getLockFilesInRemoteStore(remoteStoreEnabledIndexName, remoteStoreRepoName).length == totalShallowCopySnapshotsCount);
+        assert (getLockFilesInRemoteStore(remoteStoreEnabledIndexName, REMOTE_REPO_NAME).length == totalShallowCopySnapshotsCount);
         assert (getRepositoryData(snapshotRepoName).getSnapshotIds().size() == totalSnapshotsCount);
         // Deleting subset of shallow copy snapshots
         assertAcked(
@@ -213,7 +212,7 @@ public class DeleteSnapshotIT extends AbstractSnapshotIntegTestCase {
         totalSnapshotsCount -= tobeDeletedShallowCopySnapshotsCount;
         totalShallowCopySnapshotsCount -= tobeDeletedShallowCopySnapshotsCount;
         assert (getRepositoryData(snapshotRepoName).getSnapshotIds().size() == totalSnapshotsCount);
-        assert (getLockFilesInRemoteStore(remoteStoreEnabledIndexName, remoteStoreRepoName).length == totalShallowCopySnapshotsCount);
+        assert (getLockFilesInRemoteStore(remoteStoreEnabledIndexName, REMOTE_REPO_NAME).length == totalShallowCopySnapshotsCount);
 
         // Deleting subset of full copy snapshots
         assertAcked(
@@ -224,7 +223,7 @@ public class DeleteSnapshotIT extends AbstractSnapshotIntegTestCase {
         );
         totalSnapshotsCount -= tobeDeletedFullCopySnapshotsCount;
         assert (getRepositoryData(snapshotRepoName).getSnapshotIds().size() == totalSnapshotsCount);
-        assert (getLockFilesInRemoteStore(remoteStoreEnabledIndexName, remoteStoreRepoName).length == totalShallowCopySnapshotsCount);
+        assert (getLockFilesInRemoteStore(remoteStoreEnabledIndexName, REMOTE_REPO_NAME).length == totalShallowCopySnapshotsCount);
     }
 
     // Deleting subset of shallow and full copy snapshots as part of single delete call and then deleting all snapshots in the repo.
@@ -233,7 +232,7 @@ public class DeleteSnapshotIT extends AbstractSnapshotIntegTestCase {
         disableRepoConsistencyCheck("Remote store repository is being used in the test");
         FeatureFlagSetter.set(FeatureFlags.REMOTE_STORE);
 
-        internalCluster().startClusterManagerOnlyNode();
+        internalCluster().startClusterManagerOnlyNode(remoteStoreClusterSettings(REMOTE_REPO_NAME));
         internalCluster().startDataOnlyNode();
         final Client clusterManagerClient = internalCluster().clusterManagerClient();
         ensureStableCluster(2);
@@ -245,11 +244,10 @@ public class DeleteSnapshotIT extends AbstractSnapshotIntegTestCase {
         createIndexWithContent(testIndex);
 
         final Path remoteStoreRepoPath = randomRepoPath();
-        final String remoteStoreRepoName = "remote-store-repo-name";
-        createRepository(remoteStoreRepoName, "fs", remoteStoreRepoPath);
+        createRepository(REMOTE_REPO_NAME, "fs", remoteStoreRepoPath);
 
         final String remoteStoreEnabledIndexName = "remote-index-1";
-        final Settings remoteStoreEnabledIndexSettings = getRemoteStoreBackedIndexSettings(remoteStoreRepoName);
+        final Settings remoteStoreEnabledIndexSettings = getRemoteStoreBackedIndexSettings();
         createIndex(remoteStoreEnabledIndexName, remoteStoreEnabledIndexSettings);
         indexRandomDocs(remoteStoreEnabledIndexName, randomIntBetween(5, 10));
 
@@ -268,7 +266,7 @@ public class DeleteSnapshotIT extends AbstractSnapshotIntegTestCase {
 
         int totalSnapshotsCount = totalFullCopySnapshotsCount + totalShallowCopySnapshotsCount;
 
-        assert (getLockFilesInRemoteStore(remoteStoreEnabledIndexName, remoteStoreRepoName).length == totalShallowCopySnapshotsCount);
+        assert (getLockFilesInRemoteStore(remoteStoreEnabledIndexName, REMOTE_REPO_NAME).length == totalShallowCopySnapshotsCount);
         assert (getRepositoryData(snapshotRepoName).getSnapshotIds().size() == totalSnapshotsCount);
         // Deleting subset of shallow copy snapshots and full copy snapshots
         assertAcked(
@@ -283,12 +281,12 @@ public class DeleteSnapshotIT extends AbstractSnapshotIntegTestCase {
         totalSnapshotsCount -= (tobeDeletedShallowCopySnapshotsCount + tobeDeletedFullCopySnapshotsCount);
         totalShallowCopySnapshotsCount -= tobeDeletedShallowCopySnapshotsCount;
         assert (getRepositoryData(snapshotRepoName).getSnapshotIds().size() == totalSnapshotsCount);
-        assert (getLockFilesInRemoteStore(remoteStoreEnabledIndexName, remoteStoreRepoName).length == totalShallowCopySnapshotsCount);
+        assert (getLockFilesInRemoteStore(remoteStoreEnabledIndexName, REMOTE_REPO_NAME).length == totalShallowCopySnapshotsCount);
 
         // Deleting all the remaining snapshots
         assertAcked(clusterManagerClient.admin().cluster().prepareDeleteSnapshot(snapshotRepoName, "*").get());
         assert (getRepositoryData(snapshotRepoName).getSnapshotIds().size() == 0);
-        assert (getLockFilesInRemoteStore(remoteStoreEnabledIndexName, remoteStoreRepoName).length == 0);
+        assert (getLockFilesInRemoteStore(remoteStoreEnabledIndexName, REMOTE_REPO_NAME).length == 0);
     }
 
     private List<String> createNSnapshots(String repoName, int count) {
