+++ server/src/internalClusterTest/java/org/opensearch/snapshots/SnapshotStatusApisIT.java
@@ -112,17 +112,13 @@ public class SnapshotStatusApisIT extends AbstractSnapshotIntegTestCase {
         assertEquals(snStatus.getStats().getTime(), snapshotInfo.endTime() - snapshotInfo.startTime());
     }
 
-    public void testStatusAPICallForShallowCopySnapshot() throws Exception {
+    public void testStatusAPICallForShallowCopySnapshot() {
         disableRepoConsistencyCheck("Remote store repository is being used for the test");
         internalCluster().startClusterManagerOnlyNode();
         internalCluster().startDataOnlyNode();
 
         final String snapshotRepoName = "snapshot-repo-name";
         createRepository(snapshotRepoName, "fs", snapshotRepoSettingsForShallowCopy());
-
-        final Path remoteStoreRepoPath = randomRepoPath();
-        final String remoteStoreRepoName = "remote-store-repo-name";
-        createRepository(remoteStoreRepoName, "fs", remoteStoreRepoPath);
 
         final String indexName = "index-1";
         createIndex(indexName);
@@ -133,20 +129,8 @@ public class SnapshotStatusApisIT extends AbstractSnapshotIntegTestCase {
         }
         refresh();
 
-        final String remoteStoreEnabledIndexName = "remote-index-1";
-        final Settings remoteStoreEnabledIndexSettings = getRemoteStoreBackedIndexSettings(remoteStoreRepoName);
-        createIndex(remoteStoreEnabledIndexName, remoteStoreEnabledIndexSettings);
-        ensureGreen();
-
-        logger.info("--> indexing some data");
-        for (int i = 0; i < 100; i++) {
-            index(remoteStoreEnabledIndexName, "_doc", Integer.toString(i), "foo", "bar" + i);
-        }
-        refresh();
-
         final String snapshot = "snapshot";
         createFullSnapshot(snapshotRepoName, snapshot);
-        assert (getLockFilesInRemoteStore(remoteStoreEnabledIndexName, remoteStoreRepoName).length == 1);
 
         final SnapshotStatus snapshotStatus = getSnapshotStatus(snapshotRepoName, snapshot);
         assertThat(snapshotStatus.getState(), is(SnapshotsInProgress.State.SUCCESS));
@@ -157,14 +141,6 @@ public class SnapshotStatusApisIT extends AbstractSnapshotIntegTestCase {
         assertThat(snapshotShardState.getStats().getTotalSize(), greaterThan(0L));
         assertThat(snapshotShardState.getStats().getIncrementalFileCount(), greaterThan(0));
         assertThat(snapshotShardState.getStats().getIncrementalSize(), greaterThan(0L));
-
-        // Validating that the incremental file count and incremental file size is zero for shallow copy
-        final SnapshotIndexShardStatus shallowSnapshotShardState = stateFirstShard(snapshotStatus, remoteStoreEnabledIndexName);
-        assertThat(shallowSnapshotShardState.getStage(), is(SnapshotIndexShardStage.DONE));
-        assertThat(shallowSnapshotShardState.getStats().getTotalFileCount(), greaterThan(0));
-        assertThat(shallowSnapshotShardState.getStats().getTotalSize(), greaterThan(0L));
-        assertThat(shallowSnapshotShardState.getStats().getIncrementalFileCount(), is(0));
-        assertThat(shallowSnapshotShardState.getStats().getIncrementalSize(), is(0L));
     }
 
     public void testStatusAPICallInProgressSnapshot() throws Exception {
@@ -243,63 +219,6 @@ public class SnapshotStatusApisIT extends AbstractSnapshotIntegTestCase {
             SnapshotMissingException.class,
             () -> client().admin().cluster().prepareSnapshotStatus("test-repo").setSnapshots("test-snap").execute().actionGet()
         );
-    }
-
-    public void testStatusAPIStatsForBackToBackShallowSnapshot() throws Exception {
-        disableRepoConsistencyCheck("Remote store repository is being used for the test");
-        internalCluster().startClusterManagerOnlyNode();
-        internalCluster().startDataOnlyNode();
-
-        final String snapshotRepoName = "snapshot-repo-name";
-        createRepository(snapshotRepoName, "fs", snapshotRepoSettingsForShallowCopy());
-
-        final Path remoteStoreRepoPath = randomRepoPath();
-        final String remoteStoreRepoName = "remote-store-repo-name";
-        createRepository(remoteStoreRepoName, "fs", remoteStoreRepoPath);
-
-        final String indexName = "index-1";
-        createIndex(indexName);
-        ensureGreen();
-        logger.info("--> indexing some data");
-        for (int i = 0; i < 100; i++) {
-            index(indexName, "_doc", Integer.toString(i), "foo", "bar" + i);
-        }
-        refresh();
-
-        final String remoteStoreEnabledIndexName = "remote-index-1";
-        final Settings remoteStoreEnabledIndexSettings = getRemoteStoreBackedIndexSettings(remoteStoreRepoName);
-        createIndex(remoteStoreEnabledIndexName, remoteStoreEnabledIndexSettings);
-        ensureGreen();
-
-        logger.info("--> indexing some data");
-        for (int i = 0; i < 100; i++) {
-            index(remoteStoreEnabledIndexName, "_doc", Integer.toString(i), "foo", "bar" + i);
-        }
-        refresh();
-
-        createFullSnapshot(snapshotRepoName, "test-snap-1");
-        assert (getLockFilesInRemoteStore(remoteStoreEnabledIndexName, remoteStoreRepoName).length == 1);
-
-        SnapshotStatus snapshotStatus = getSnapshotStatus(snapshotRepoName, "test-snap-1");
-        assertThat(snapshotStatus.getState(), is(SnapshotsInProgress.State.SUCCESS));
-
-        SnapshotIndexShardStatus shallowSnapshotShardState = stateFirstShard(snapshotStatus, remoteStoreEnabledIndexName);
-        assertThat(shallowSnapshotShardState.getStage(), is(SnapshotIndexShardStage.DONE));
-        final int totalFileCount = shallowSnapshotShardState.getStats().getTotalFileCount();
-        final long totalSize = shallowSnapshotShardState.getStats().getTotalSize();
-        final int incrementalFileCount = shallowSnapshotShardState.getStats().getIncrementalFileCount();
-        final long incrementalSize = shallowSnapshotShardState.getStats().getIncrementalSize();
-
-        createFullSnapshot(snapshotRepoName, "test-snap-2");
-        assert (getLockFilesInRemoteStore(remoteStoreEnabledIndexName, remoteStoreRepoName).length == 2);
-
-        snapshotStatus = getSnapshotStatus(snapshotRepoName, "test-snap-2");
-        assertThat(snapshotStatus.getState(), is(SnapshotsInProgress.State.SUCCESS));
-        shallowSnapshotShardState = stateFirstShard(snapshotStatus, remoteStoreEnabledIndexName);
-        assertThat(shallowSnapshotShardState.getStats().getTotalFileCount(), equalTo(totalFileCount));
-        assertThat(shallowSnapshotShardState.getStats().getTotalSize(), equalTo(totalSize));
-        assertThat(shallowSnapshotShardState.getStats().getIncrementalFileCount(), equalTo(incrementalFileCount));
-        assertThat(shallowSnapshotShardState.getStats().getIncrementalSize(), equalTo(incrementalSize));
     }
 
     public void testGetSnapshotsWithoutIndices() throws Exception {
@@ -441,16 +360,11 @@ public class SnapshotStatusApisIT extends AbstractSnapshotIntegTestCase {
     }
 
     public void testStatusAPICallInProgressShallowSnapshot() throws Exception {
-        disableRepoConsistencyCheck("Remote store repository is being used for the test");
         internalCluster().startClusterManagerOnlyNode();
         internalCluster().startDataOnlyNode();
 
         final String snapshotRepoName = "snapshot-repo-name";
         createRepository(snapshotRepoName, "mock", snapshotRepoSettingsForShallowCopy().put("block_on_data", true));
-
-        final Path remoteStoreRepoPath = randomRepoPath();
-        final String remoteStoreRepoName = "remote-store-repo-name";
-        createRepository(remoteStoreRepoName, "mock", remoteStoreRepoPath);
 
         final String indexName = "index-1";
         createIndex(indexName);
@@ -458,17 +372,6 @@ public class SnapshotStatusApisIT extends AbstractSnapshotIntegTestCase {
         logger.info("--> indexing some data");
         for (int i = 0; i < 100; i++) {
             index(indexName, "_doc", Integer.toString(i), "foo", "bar" + i);
-        }
-        refresh();
-
-        final String remoteStoreEnabledIndexName = "remote-index-1";
-        final Settings remoteStoreEnabledIndexSettings = getRemoteStoreBackedIndexSettings(remoteStoreRepoName);
-        createIndex(remoteStoreEnabledIndexName, remoteStoreEnabledIndexSettings);
-        ensureGreen();
-
-        logger.info("--> indexing some data");
-        for (int i = 0; i < 100; i++) {
-            index(remoteStoreEnabledIndexName, "_doc", Integer.toString(i), "foo", "bar" + i);
         }
         refresh();
